pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "brief-breakdown"
version = "0.1.0"
description = "Turn agency project briefs into structured task plans, with automated evals."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
brief-breakdown = "brief_breakdown.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
